pyproject.toml
[project]
name = "ring-tsfast"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "numpy",
    "qmt",
    "tsfast",
    "torch-optimizer",
    "pytorch-optimizer>=3.10.0",
]

[project.optional-dependencies]
jax = ["imt-ring", "fire", "wandb"]
imt = ["imt-imt", "imt-diodem", "onnxruntime-gpu", "tensorrt-cu12"]

[tool.uv]
exclude-dependencies = ["onnxruntime"]

[tool.uv.sources]
tsfast = { path = "../tsfast", editable = true }
imt-imt = { git = "https://github.com/simon-bachhuber/imt.git" }

[tool.setuptools]
py-modules = []
